Simplify an example
In this example, we're making sure that `view` is available in hooks.
There's no big difference in globally defined hooks and local hooks, so
it's not necessary to define one on the `configuration`, and thus
`with_isolated_config` becomes unnecessary as well.

`group.run` accepts a reporter, but it defaults to a NullReporter
anyway.

In this case, it's not really necessary to spy on the reporter and make
sure no errors happened - we increment only after we've referenced
`view`, if there would be an error in the hook, the increment wouldn't
run as well.